+++ apps/web/package.json
@@ -11,7 +11,6 @@
 		"@fontsource/inter": "^4.5.11",
 		"@rspc/client": "^0.0.6",
 		"@sd/client": "workspace:*",
-		"@sd/core": "workspace:*",
 		"@sd/interface": "workspace:*",
 		"@sd/ui": "workspace:*",
 		"@tanstack/react-query": "^4.0.10",
