+++ package.json
@@ -14,7 +14,6 @@
 		"landing": "pnpm --filter @sd/landing -- ",
 		"ui": "pnpm --filter @sd/ui -- ",
 		"interface": "pnpm --filter @sd/interface -- ",
-		"core": "pnpm --filter @sd/core -- ",
 		"docs": "pnpm --filter @sd/docs -- ",
 		"client": "pnpm --filter @sd/client -- ",
 		"server": "pnpm --filter @sd/server -- ",
