+++ packages/client/package.json
@@ -20,7 +20,6 @@
 		"@rspc/client": "^0.0.6",
 		"@rspc/react": "^0.0.6",
 		"@sd/config": "workspace:*",
-		"@sd/core": "workspace:*",
 		"@sd/interface": "workspace:*",
 		"@tanstack/react-query": "^4.0.10",
 		"eventemitter3": "^4.0.7",
