+++ packages/client/src/index.ts
@@ -1,5 +1,5 @@
 export * from './stores';
 export * from './context';
 export * from './rspc';
+export * from './core';
 export * from './hooks';
-export type { Operations } from '@sd/core';
