+++ packages/interface/package.json
@@ -27,7 +27,6 @@
 		"@radix-ui/react-tooltip": "^1.0.0",
 		"@sd/assets": "workspace:*",
 		"@sd/client": "workspace:*",
-		"@sd/core": "workspace:*",
 		"@sd/ui": "workspace:*",
 		"@tailwindcss/forms": "^0.5.2",
 		"@tanstack/react-query": "^4.2.3",
