+++ packages/interface/src/screens/settings/library/TagsSettings.tsx
@@ -1,3 +1,12 @@
+import { TrashIcon } from '@heroicons/react/24/outline';
+import { useLibraryMutation, useLibraryQuery } from '@sd/client';
+import { TagUpdateArgs } from '@sd/client';
+import { Button, Input } from '@sd/ui';
+import clsx from 'clsx';
+import { useCallback, useEffect, useMemo, useState } from 'react';
+import { Controller, useForm } from 'react-hook-form';
+import { useDebounce } from 'rooks';
+
 import Card from '../../../components/layout/Card';
 import Dialog from '../../../components/layout/Dialog';
 import { Toggle } from '../../../components/primitive';
@@ -5,14 +14,6 @@ import { InputContainer } from '../../../components/primitive/InputContainer';
 import { PopoverPicker } from '../../../components/primitive/PopoverPicker';
 import { SettingsContainer } from '../../../components/settings/SettingsContainer';
 import { SettingsHeader } from '../../../components/settings/SettingsHeader';
-import { TrashIcon } from '@heroicons/react/24/outline';
-import { useLibraryMutation, useLibraryQuery } from '@sd/client';
-import { TagUpdateArgs } from '@sd/core';
-import { Button, Input } from '@sd/ui';
-import clsx from 'clsx';
-import { useCallback, useEffect, useMemo, useState } from 'react';
-import { Controller, useForm } from 'react-hook-form';
-import { useDebounce } from 'rooks';
 
 export default function TagsSettings() {
 	const [openCreateModal, setOpenCreateModal] = useState(false);
